requirements.txt
# Web
Django==4.1.1
djangorestframework==3.14.0

# Database
psycopg2-binary==2.9.3
